refactor: move HasApiResponse to trait folder
