+++ app/Http/Middleware/RoleMiddleware.php
@@ -3,9 +3,8 @@
 namespace App\Http\Middleware;
 
 use App\Enums\Role;
-use App\Http\Controllers\Concerns\HasApiResponse;
+use App\Traits\HasApiResponse;
 use Closure;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use RuntimeException;
 use Symfony\Component\HttpFoundation\Response;
@@ -21,6 +20,20 @@ class RoleMiddleware
      */
     public function handle(Request $request, Closure $next, ...$roles): Response
     {
+        if (! $request->user()
+            || ! $this->userHasRole($request->user(), $roles)
+        ) {
+            return $this->errorResponse(
+                message: 'You do not have the required role to access this resource.',
+                statusCode: 403,
+            );
+        }
+
+        return $next($request);
+    }
+
+    protected function getAllowedRoles(array $roles)
+    {
         $allowedRoles = [];
         foreach ($roles as $role) {
             $roleEnum = Role::tryFrom($role);
@@ -30,11 +43,11 @@ class RoleMiddleware
             $allowedRoles[] = $roleEnum;
         }
 
-        throw_if(
-            ! $request->user() || ! in_array($request->user()->role, $allowedRoles),
-            new AuthorizationException('You do not have the required role to access this resource.')
-        );
+        return $allowedRoles;
+    }
 
-        return $next($request);
+    protected function userHasRole($user, $roles)
+    {
+        return in_array($user->role, $this->getAllowedRoles($roles));
     }
 }
